Update Todos show to screen
Todos now show on screen and will add the next todo to the bottom
of the list. The plan is to change the order only when a Todo is updated.
If the todo is marked as complete it should be shifted to the bottom,
top of the stack at the bottom should be the most recent Todo. Will
be adding a filter as well to show only completed Todos.